+++ CHANGELOG.md
@@ -63,6 +63,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Supported by: `demonstrative`, `instructions`, `header_info`
 - Title `getActualValue()`
 - ShippingFile `date()`
+- Optional 'R' segment in CNAB240 ShippingFile
 
 ### Changed
 - Update dependencies